Move BooleanGuard to logic package
