+++ core/src/main/kotlin/io/github/kerelape/reikai/logic/BooleanGuard.kt
@@ -1,4 +1,4 @@
-package io.github.kerelape.reikai.guards
+package io.github.kerelape.reikai.logic
 
 import io.github.kerelape.reikai.core.Entity
 import io.github.kerelape.reikai.core.EntityWrap
